authorization/urls.py
from django.urls import path
from . import views
from rest_framework_simplejwt.views import TokenObtainPairView


urlpatterns = [
    path('auth/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('register/', views.RegisterView.as_view(), name='user_register'),
    path('update/', views.UpdateParamsView.as_view(), name='user_update'),
]
