requirements.txt
Django>=3.1,<4.0
psycopg2-binary==2.8.6
djangorestframework
django-cors-headers
djangorestframework_simplejwt
pytest
pytest-django 
